requirements/base.txt
pytz==2018.9  # https://github.com/stub42/pytz
python-slugify==2.0.1  # https://github.com/un33k/python-slugify
Pillow==5.4.1  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
celery[redis]==4.3.0
django-celery-results
drf-base64==2.0
vine==1.3.0
django-xworkflows==1.0.0

# Django
# ------------------------------------------------------------------------------
django==2.2.6          # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-des==2.4.1 # https://github.com/jamiecounsell/django-des
django-cors-headers==3.0.2   # https://pypi.org/project/django-cors-headers/

# Django REST Framework
djangorestframework==3.9.1  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
djangorestframework-jwt==1.11.0 # https://getblimp.github.io/django-rest-framework-jwt/
django-rest-swagger==2.2.0 # https://github.com/marcgibbons/django-rest-swagger
